refactor: Rename `misc` to `utils`
Align with Gutenberg project practices.
